Finally, 8 puzzle problem works!
The issue was with python not allowing pass by value for lists (as a matter of fact there is no pass by value in python), hence everytime a list was passed as an argument, the data was manipulated on the original list instead of a copy of that. Hence used deep copy to create a copy and then operations were performed on that. It works!!